REFACTOR (Drawer Modal): Transferred Drawer Modal to its own file
